requirements.txt
google-generativeai
python-dotenv
streamlit
langchain
PyPDF2
faiss-cpu
langchain_google_genai